bintray-package: Download files in the `in` script
Only files matching the filter in the remote path are downloaded. They
are placed in the specified local path.

The layout is reproduced locally. In other words, if the remote file is
in a subdirectory of the remote path, it will be placed in the same
subdirectory of the local path.

[#142215799]